split off queryDbRow from ZhangScoreFunctions into separate class DbFunctions
